+++ src/features/trip/components/PreviewImageDialog.tsx
